[Kernel] Allow kernel to write Type Widening enable (#4517)
#### Which Delta project/connector is this regarding?
<!--
Please add the component selected below to the beginning of the pull
request title
For example: [Spark] Title of my pull request
-->

- [ ] Spark
- [ ] Standalone
- [ ] Flink
- [x] Kernel
- [ ] Other (fill in here)

## Description

Currently having this flag return 'false' prevents kernel from writing
any table where this is feature is present. this is the first change for
#4491 .

## How was this patch tested?

N/A flag flip

## Does this PR introduce _any_ user-facing changes?

Yes, kernel will now be able to write the table with this table feature
set.